+++ lib/tinderfridge/devices/bricklet_voltage_current_v2.rb
@@ -1,0 +1,18 @@
+module Tinkerforge
+
+  class BrickletVoltageCurrentV2
+
+    # Returns the device's state.
+    def state
+      super.merge(
+        'voltage'       => get_voltage,
+        'current'       => get_current,
+        'power'         => get_power,
+        'configuration' => get_configuration,
+        'calibration'   => get_calibration,
+      )
+    end
+
+  end
+
+end