+++ History.md
@@ -2,6 +2,12 @@
 
 # History
 
+## Unreleased
+
+* Bricklet classes:
+    * `BrickletVoltageCurrentV2`:
+        * `#state` includes voltage, current, power, configuration, and calibration.
+
 ## 0.5.0 (12 March 2021)
 [GitHub](https://github.com/lllisteu/tinderfridge/releases/tag/v0.5.0) • [RubyGems](https://rubygems.org/gems/tinderfridge/versions/0.5.0) • [Documentation](https://www.rubydoc.info/gems/tinderfridge/0.5.0)